+++ R/RcppExports.R
@@ -19,6 +19,12 @@ lfdistance_new <- function(L1, D1, L2, D2, timestamp) {
 
 #' @keywords internal
 #' @noRd
+lfdistance_new_faster <- function(vecs, vals, timestamps) {
+    .Call('_NetworkDistance_lfdistance_new_faster', PACKAGE = 'NetworkDistance', vecs, vals, timestamps)
+}
+
+#' @keywords internal
+#' @noRd
 cpp_gdd <- function(vecs, vals, timestamps) {
     .Call('_NetworkDistance_cpp_gdd', PACKAGE = 'NetworkDistance', vecs, vals, timestamps)
 }