+++ NEWS.md
@@ -1,5 +1,8 @@
-## News for Package NetworkDistance
-## Changes in version 0.3.1
+# News for Package NetworkDistance
+### Changes in version 0.3.2
+  * faster implementation of `nd.nfd`
+  
+### Changes in version 0.3.1
   * Error fixed in `nd.him` (Franck Lejzerowicz at UCSD).
   * Initialize the following documentation:
     - NEWS for keeping record of updates.